requirements.txt
autopep8==2.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.0
soupsieve==2.3.2.post1
tomli==2.0.1
Werkzeug==2.2.2
requests==2.28.2
